backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pyjwt==2.8.0
groq==0.9.0
textblob==0.18.0
python-dotenv==1.0.1
nltk==3.8.1
pymongo==4.7.3